requirements.txt
Django~=4.2.29
pyTelegramBotAPI~=4.32.0
djangorestframework~=3.16.1